[test] apply the new test style to common-meta-types
