+++ common/meta/types/Cargo.toml
@@ -6,6 +6,10 @@ license = "Apache-2.0"
 publish = false
 edition = "2021"
 
+[lib]
+doctest = false
+test = false
+
 [dependencies]
 common-datavalues = {path = "../../datavalues"}
 common-exception = {path = "../../exception"}
