+++ common/meta/types/tests/it/user_quota.rs
@@ -13,23 +13,14 @@
 // limitations under the License.
 
 use common_exception::exception::Result;
-
-use crate::UserPrivilege;
-use crate::UserPrivilegeType;
+use common_meta_types::UserQuota;
 
 #[test]
-fn test_user_privilege() -> Result<()> {
-    let mut privileges = UserPrivilege::empty();
-    let r = privileges.has_privilege(UserPrivilegeType::Set);
-    assert!(!r);
-
-    privileges.set_privilege(UserPrivilegeType::Set);
-    let r = privileges.has_privilege(UserPrivilegeType::Set);
-    assert!(r);
-
-    privileges.set_all_privileges();
-    let r = privileges.has_privilege(UserPrivilegeType::Create);
-    assert!(r);
+fn test_user_quota() -> Result<()> {
+    let quota = UserQuota::no_limit();
+    assert_eq!(quota.max_cpu, 0);
+    assert_eq!(quota.max_memory_in_bytes, 0);
+    assert_eq!(quota.max_storage_in_bytes, 0);
 
     Ok(())
 }
